--- fees/avalon-finance-usda/constants.ts
@@ -1,28 +1,0 @@
-import ADDRESSES from '../../helpers/coreAssets.json'
-
-export const TOKEN_DETAILS = {
-	[ADDRESSES.ethereum.USDa]: {
-		coingeckoId: 'usda-2',
-		decimals: 18,
-	},
-	[ADDRESSES.bsc.USDa]: {
-		coingeckoId: 'usda-2',
-		decimals: 18,
-	},
-	[ADDRESSES.mantle.USDa]: {
-		coingeckoId: 'usda-2',
-		decimals: 18,
-	},
-	[ADDRESSES.sonic.USDa]: {
-		coingeckoId: 'usda-2',
-		decimals: 18,
-	},
-	[ADDRESSES.berachain.USDa]: {
-		coingeckoId: 'usda-2',
-		decimals: 18,
-	},
-	[ADDRESSES.klaytn.USDa]: {
-		coingeckoId: 'usda-2',
-		decimals: 18,
-	},
-}
