+++ helpers/coreAssets.json
@@ -68,8 +68,7 @@
     "BTCN": "0x386E7A3a0c0919c9d53c3b04FF67E73Ff9e45Fb6",
     "POL": "0x455e53cbb86018ac2b8092fdcd39d8444affc3f6",
     "WSOL": "0xD31a59c85aE9D8edEFeC411D448f90841571b89c",
-    "WAVES": "0x0Fb765ddBD4d26AC524AA5990B0643D0Ab6Ac2fE",
-    "USDa": "0x8A60E489004Ca22d775C5F2c657598278d17D9c2"
+    "WAVES": "0x0Fb765ddBD4d26AC524AA5990B0643D0Ab6Ac2fE"
   },
   "fantom": {
     "WFTM": "0x21be370d5312f44cb42ce377bc9b8a0cef1a4c83",
@@ -150,8 +149,7 @@
     "aBNBb": "0xbb1aa6e59e5163d8722a122cd66eba614b59df0d",
     "TUSD": "0x40af3827f39d0eacbf4a168f8d4ee67c121d11c9",
     "WBTC": "0x0555E30da8f98308EdB960aa94C0Db47230d2B9c",
-    "weETH": "0x04C0599Ae5A44757c0af6F9eC3b93da8976c150A",
-    "USDa": "0x9356086146be5158E98aD827E21b5cF944699894"
+    "weETH": "0x04C0599Ae5A44757c0af6F9eC3b93da8976c150A"
   },
   "polygon": {
     "WMATIC": "0xfd28c7cea3c50a060cb4c0059e453c6d4dd9829d",
@@ -510,8 +508,7 @@
     "USDK": "0xd2137fdf10bd9e4e850c17539eb24cfe28777753",
     "NPT": "0xe06597d02a2c3aa7a9708de2cfa587b128bd3815",
     "USDT_1": "0xd077a400968890eacc75cdc901f0356c943e4fdb",
-    "stKaia": "0x42952B873ed6f7f0A7E4992E2a9818E3A9001995",
-    "USDa": "0xdc3cf1961b08da169b078f7df6f26676bf6a4ff6"
+    "stKaia": "0x42952B873ed6f7f0A7E4992E2a9818E3A9001995"
   },
   "nova": {
     "NUSD": "0x1f5396f254ee25377a5c1b9c6bff5f44e9294fff"
@@ -1565,8 +1562,7 @@
     "FBTC": "0xC96dE26018A54D51c097160568752c4E3BD6C364",
     "AUSD": "0x00000000eFE302BEAA2b3e6e1b18d08D69a9012a",
     "USDe": "0x5d3a1Ff2b6BAb83b63cd9AD0787074081a52ef34",
-    "sUSDe": "0x211Cc4DD073734dA055fbF44a2b4667d5E5fE5d2",
-    "USDa": "0x075df695b8E7f4361FA7F8c1426C63f11B06e326"
+    "sUSDe": "0x211Cc4DD073734dA055fbF44a2b4667d5E5fE5d2"
   },
   "linea": {
     "WETH": "0xe5D7C2a44FfDDf6b295A15c148167daaAf5Cf34f",
@@ -2207,8 +2203,7 @@
     "scUSD": "0xd3DCe716f3eF535C5Ff8d041c1A41C3bd89b97aE",
     "USDT": "0x6047828dc181963ba44974801ff68e538da5eaf9",
     "LBTC": "0xecAc9C5F704e954931349Da37F60E39f515c11c1",
-    "WBTC": "0x3C1BCa5a656e69edCD0D4E36BEbb3FcDAcA60Cf1",
-    "USDa": "0xff12470a969dd362eb6595ffb44c82c959fe9acc"
+    "WBTC": "0x3C1BCa5a656e69edCD0D4E36BEbb3FcDAcA60Cf1"
   },
   "vinu": {
     "USDT": "0xC0264277fcCa5FCfabd41a8bC01c1FcAF8383E41",
@@ -2370,8 +2365,7 @@
     "USDC": "0x549943e04f40284185054145c6E4e9568C1D3241",
     "HONEY": "0xFCBD14DC51f0A4d49d5E53C2E0950e0bC26d0Dce",
     "WETH": "0x2F6F07CDcf3588944Bf4C42aC74ff24bF56e7590",
-    "WFBTC": "0xA18717e22dfAbB36f0242259df4e9E563b6Ec70a",
-    "USDa": "0xff12470a969dd362eb6595ffb44c82c959fe9acc"
+    "WFBTC": "0xA18717e22dfAbB36f0242259df4e9E563b6Ec70a"
   },
   "stratis": {
     "WSTRAX": "0xeA705D2DbD8DE7Dc70Db7B531D0F620d9CeE9d18"
